feat: 결제 생성 및 승인 흐름 보강

+++ src/main/java/com/v_payment/pay/payment/infra/SuccessResult.java
@@ -10,7 +10,7 @@ public record SuccessResult(
         String orderId,
         String paymentKey,
         String status,
-        BigDecimal totalAmount,
+        Long totalAmount,
         LocalDateTime approvedAt,
         Receipt receipt
 ) implements Result {
